Do not enqueue unblocked commands from host task thread
enqueueing commands from the host task thread causes a race condition for very
quickly executing lambdas. This is because commands are enqueued both from
the host thread, and then also from the HT thread if the command is blocked
by the lambda executing on HT thread.

If the command blocked by HT A starts and finishes executing and is cleaned up
by the main thread before thread A enqueues the new command then the new
enqueue will be using deallocated data. This could cause some non determinism
or segfaulting.